Normalize.jl

Normalize.jl is a small package that can consume nested data structures like dictionaries of dictionaries or structs of structs and produce a normalized, Tables.jl-compliant NamedTuple. It can be used with JSON3.jl, XMLDict.jl, and other packages that parse file formats which are structured as denormalized data.

Normalize.jl

Getting Started

Install

Basic Usage

Configuring Options

Keyword Arguments

Using ColumnDefintions

API

Getting Started

Install

julia> using Pkg
julia> Pkg.add(url="https://github.com/mrufsvold/Normalize.jl")     Cloning git-repo `https://github.com/mrufsvold/Normalize.jl`
    Updating git-repo `https://github.com/mrufsvold/Normalize.jl`
    Updating registry at `~/.julia/registries/General`
    Updating git-repo `https://github.com/JuliaRegistries/General.git`
   Resolving package versions...
    Updating `~/work/Normalize.jl/Normalize.jl/docs/Project.toml`
  [8a7d223a] ~ Normalize v0.1.0 `~/work/Normalize.jl/Normalize.jl` ⇒ v0.1.0 `https://github.com/mrufsvold/Normalize.jl#main`
    Updating `~/work/Normalize.jl/Normalize.jl/docs/Manifest.toml`
  [8a7d223a] ~ Normalize v0.1.0 `~/work/Normalize.jl/Normalize.jl` ⇒ v0.1.0 `https://github.com/mrufsvold/Normalize.jl#main`
Precompiling project...
  ✓ Normalize
  1 dependency successfully precompiled in 1 seconds. 31 already precompiled.
  1 dependency precompiled but a different version is currently loaded. Restart julia to access the new version

Basic Usage

Normalize provides a single function normalize to flatten out nested data.

using JSON3
using DataFrames

message = JSON3.read("""
    {
        "a" : [
            {"b" : 1, "c" : 2},
            {"b" : 2},
            {"b" : [3, 4], "c" : 1},
            {"b" : []}
        ],
        "d" : 4
    }
    """
)

normalize(message) |> DataFrame

4×3 DataFrame

Row	d	a_b	a_c

Int64	Union…?	Int64?

1	4	1	2

2	4	2	missing

3	4	[3, 4]	1

4	4	missing	missing

Configuring Options

While normalize can produce a Table out of the box, it is often useful to configure some options in how it handles the normalization process. Normalize.jl offers two ways to set these configurations. You can set them globally with kwargs or exercise finer control with per-column configurations.

Keyword Arguments

Parameter	Description

flatten_arrays::Bool	When a leaf node is an array, should the values be flattened into separate rows or treated as a single value. Default: true

default_value::Any	When a certain key exists in one branch, but not another, what value should be used to fill missing. Default: missing

pool_arrays::Bool	When collecting vectors for columns, choose whether to use PooledArrays instead of Base.Vector

column_names::Dict{Vector{Symbol}, Symbol}	Provide a mapping of key/fieldname paths to replaced column names

message = JSON3.read("""
    {
        "a" : [
            {"b" : 1, "c" : 2},
            {"b" : 2},
            {"b" : [3, 4], "c" : 1},
            {"b" : []}
        ],
        "d" : 4
    }
    """
)

name_map = Dict([:a, :b] => :Column_B)

normalize(message; flatten_arrays=true, default_value="no value", pool_arrays=true, column_names=name_map) |> DataFrame

5×3 DataFrame

Row	d	Column_B	a_c

Int64	Union…	Union…

1	4	1	2

2	4	2	no value

3	4	3	1

4	4	4	1

5	4	no value	no value

Using ColumnDefintions

Instead of setting the configurations for the whole dataset, you can use a Vector{ColumnDefinition} to control how each column is handled. ColumnDefinition has the added benefit of allowing you to ignore certain fields from the input.

message = JSON3.read("""
    {
        "a" : [
            {"b" : 1, "c" : 2},
            {"b" : 2},
            {"b" : [3, 4], "c" : 1},
            {"b" : []}
        ],
        "d" : 4
    }
    """
)
column_defs = [
    ColumnDefinition([:d]; column_name = :ColumnD),
    ColumnDefinition([:a, :b]; flatten_arrays=true),
    ColumnDefinition([:e, :f]; column_name = :MissingColumn, default_value="Missing branch")
]

normalize(message, column_defs) |> DataFrame

5×3 DataFrame

Row	ColumnD	a_b	MissingColumn

Int64	Int64?	String

1	4	1	Missing branch

2	4	2	Missing branch

3	4	3	Missing branch

4	4	4	Missing branch

5	4	missing	Missing branch

The only difference in the kwargs API here is that column_names is column_name and accepts a single Symbol.

API

Normalize.normalize — Method

normalize(data; flatten_arrays = false, default_value = missing, pool_arrays = false, column_names=Dict{Vector{Symbol}, Symbol}())

Take a nested data structure, data and convert it into a Table

Args

data: Any nested data structure (struct of structs or Dict of Dicts) or an array of such data structures

Keyword Args

flatten_arrays: When a leaf node is an array, should the values be flattened into separate rows or treated as a single value. Default: true

default_value: When a certain key exists in one branch, but not another, what value should be used to fill missing. Default: missing

pool_arrays: When collecting vectors for columns, choose whether to use PooledArrays instead of Base.Vector. Default: false (use Vector)

column_names::Dict{Vector{Symbol}, Symbol}: Provide a mapping of key/fieldname paths to replaced column names

Returns

::NamedTuple: A Tables.jl compliant Tuple of Vectors

Normalize.normalize — Method

normalize(data, column_defs::Vector{ColumnDefinition})

Take a nested data structure, data and convert it into a Table based on configurations passed for each column.

Args

data: Any nested data structure (struct of structs or Dict of Dicts) or an array of such data structures

column_defs::Vector{ColumnDefinition}: A ColumnDefinition for each column to be extracted from the data

Returns

::NamedTuple: A Tables.jl compliant Tuple of Vectors

Normalize.ColumnDefinition — Type

ColumnDefinition(field_path; column_name=nothing, flatten_arrays=false, default_value=missing, pool_arrays=false)

Args

field_path: Vector of keys/fieldnames that constitute a path from the top of the data to the values to extract for the column

Keyword Args

column_name::Symbol: A name for the resulting column. If nothing, defaults to joining the fieldpath with snakecase_format.

flatten_arrays: When a leaf node is an array, should the values be flattened into separate rows or treated as a single value. Default: true

default_value: When the field_path keys do not exist on one or more branches, fill with this value. Default: missing

pool_arrays: When collecting values for this column, choose whether to use PooledArrays instead of Base.Vector. Default: false (use Vector)

Returns

::ColumnDefinition