ExpandNestedData.jl

ExpandNestedData.jl is a small package that can consume nested data structures like dictionaries of dictionaries or structs of structs and produce a normalized, Tables.jl-compliant NamedTuple. It can be used with JSON3.jl, XMLDict.jl, and other packages that parse file formats which are structured as denormalized data.

ExpandNestedData.jl

Getting Started

Install

Basic Usage

Configuring Options

Keyword Arguments

Using ColumnDefintions

ColumnStyles

API

Getting Started

Install

julia> using Pkg
julia> Pkg.add(url="https://github.com/mrufsvold/ExpandNestedData.jl")     Cloning git-repo `https://github.com/mrufsvold/ExpandNestedData.jl`
    Updating git-repo `https://github.com/mrufsvold/ExpandNestedData.jl`
    Updating registry at `~/.julia/registries/General`
    Updating git-repo `https://github.com/JuliaRegistries/General.git`
   Resolving package versions...
    Updating `~/work/ExpandNestedData.jl/ExpandNestedData.jl/docs/Project.toml`
  [8a7d223a] ~ ExpandNestedData v0.1.0 `~/work/ExpandNestedData.jl/ExpandNestedData.jl` ⇒ v0.1.0 `https://github.com/mrufsvold/ExpandNestedData.jl#main`
    Updating `~/work/ExpandNestedData.jl/ExpandNestedData.jl/docs/Manifest.toml`
  [8a7d223a] ~ ExpandNestedData v0.1.0 `~/work/ExpandNestedData.jl/ExpandNestedData.jl` ⇒ v0.1.0 `https://github.com/mrufsvold/ExpandNestedData.jl#main`
Precompiling project...
  ✓ ExpandNestedData
  1 dependency successfully precompiled in 1 seconds. 36 already precompiled.
  1 dependency precompiled but a different version is currently loaded. Restart julia to access the new version

Basic Usage

ExpandNestedData provides a single function expand to flatten out nested data.

using JSON3
using DataFrames

message = JSON3.read("""
    {
        "a" : [
            {"b" : 1, "c" : 2},
            {"b" : 2},
            {"b" : [3, 4], "c" : 1},
            {"b" : []}
        ],
        "d" : 4
    }
    """
)

expand(message) |> DataFrame

4×3 DataFrame

Row	d	a_b	a_c

Int64	Union…	Int64?

1	4	1	2

2	4	2	missing

3	4	[3, 4]	1

4	4	Union{}[]	missing

Configuring Options

While exapnd can produce a Table out of the box, it is often useful to configure some options in how it handles the normalization process. ExpandNestedData.jl offers two ways to set these configurations. You can set them globally with kwargs or exercise finer control with per-column configurations.

Keyword Arguments

Parameter	Description

flatten_arrays::Bool	When a leaf node is an array, should the values be flattened into separate rows or treated as a single value. Default: true

default_value::Any	When a certain key exists in one branch, but not another, what value should be used to fill missing. Default: missing

pool_arrays::Bool	When collecting vectors for columns, choose whether to use PooledArrays instead of Base.Vector

column_names::Dict{Vector{Symbol}, Symbol}	Provide a mapping of key/fieldname paths to replaced column names

column_style::T<:ExpandNestedData.ColumnStyle	Chose returned column style from nested_columns or flat_columns. If nested, columnnames are ignored and a TypedTables.Table is returned for which the columns are nested in the same structure as the source data. Default: `flatcolumns`

name_map = Dict([:a, :b] => :Column_B)
expand(message; flatten_arrays=true, default_value="no value", pool_arrays=true, column_names=name_map) |> DataFrame

5×3 DataFrame

Row	d	Column_B	a_c

Int64	Union…	Union…

1	4	1	2

2	4	2	no value

3	4	3	1

4	4	4	1

5	4	no value	no value

Using ColumnDefintions

Instead of setting the configurations for the whole dataset, you can use a Vector{ColumnDefinition} to control how each column is handled. ColumnDefinition has the added benefit of allowing you to ignore certain fields from the input.

column_defs = [
    ColumnDefinition([:d]; column_name = :ColumnD),
    ColumnDefinition([:a, :b]; flatten_arrays=true),
    ColumnDefinition([:e, :f]; column_name = :MissingColumn, default_value="Missing branch")
]

expand(message, column_defs) |> DataFrame

5×3 DataFrame

Row	MissingColumn	ColumnD	a_b

String	Int64	Int64?

1	Missing branch	4	1

2	Missing branch	4	2

3	Missing branch	4	3

4	Missing branch	4	4

5	Missing branch	4	missing

The only difference in the kwargs API here is that column_names is column_name and accepts a single Symbol.

ColumnStyles

In the examples above, we've used flat_columns style. However, we can also maintain the nesting heirarchy of the source data.

using TypedTables

tbl = expand(message; column_style = nested_columns)
tbl.a.b[1] == 1 # true
# `rows(tbl)` returns a nested NamedTuple
tbl |> rows |> first

(d = 4, a = NamedTuple{(:b, :c), Tuple{Union{Int64, Vector{Any}, Vector{Int64}}, Union{Missing, Int64}}}((1, 2)))

API

ExpandNestedData.expand — Method

expand(data; flatten_arrays = false, default_value = missing, pool_arrays = false, column_names=Dict{Vector{Symbol}, Symbol}())

Take a nested data structure, data and convert it into a Table

Args

data: Any nested data structure (struct of structs or Dict of Dicts) or an array of such data structures

Keyword Args

flatten_arrays: When a leaf node is an array, should the values be flattened into separate rows or treated as a single value. Default: true

default_value: When a certain key exists in one branch, but not another, what value should be used to fill missing. Default: missing

pool_arrays: When collecting vectors for columns, choose whether to use PooledArrays instead of Base.Vector. Default: false (use Vector)

column_names::Dict{Vector{Symbol}, Symbol}: Provide a mapping of key/fieldname paths to replaced column names

column_style: Chose returned column style from nested_columns or flat_columns. If nested, columnnames are ignored and a TypedTables.Table is returned for which the columns are nested in the same structure as the source data. Default: `flatcolumns`

Returns

::NamedTuple: A Tables.jl compliant Tuple of Vectors

ExpandNestedData.expand — Method

expand(data, column_defs::Vector{ColumnDefinition})

Take a nested data structure, data and convert it into a Table based on configurations passed for each column.

Args

data: Any nested data structure (struct of structs or Dict of Dicts) or an array of such data structures

column_defs::Vector{ColumnDefinition}: A ColumnDefinition for each column to be extracted from the data

column_style: Chose returned column style from nested_columns or flat_columns. If nested, columnnames are ignored and a TypedTables.Table is returned for which the columns are nested in the same structure as the source data. Default: `flatcolumns`

Returns

::NamedTuple: A Tables.jl compliant Tuple of Vectors

ExpandNestedData.ColumnDefinition — Type

ColumnDefinition provides a mechanism for specifying details for extracting data from a nested data source